requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
python-dotenv
reportlab
pydantic
langchain-groq